extend enums for the next house

+++ src/main/java/com/example/mingkhparser/models/coldwatersystem/ColdWaterSystem.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class ColdWaterSystem {
-    private Integer physicalDeterioration;
+    private Double physicalDeterioration;
 
     private NetworkMaterial networkMaterial;
 }
