+++ src/main/java/com/example/mingkhparser/models/hotwatersupplysystem/HotWaterSupplySystem.java
@@ -1,9 +1,5 @@
 package com.example.mingkhparser.models.hotwatersupplysystem;
 
-import com.example.mingkhparser.models.hotwatersupplysystem.HotWaterSystemType;
-import com.example.mingkhparser.models.hotwatersupplysystem.NetworkMaterial;
-import com.example.mingkhparser.models.hotwatersupplysystem.NetworkThermalInsulationMaterial;
-import com.example.mingkhparser.models.hotwatersupplysystem.RisersMaterial;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -13,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class HotWaterSupplySystem {
     private HotWaterSystemType hotWaterSystemType;
-    private Integer physicalDeterioration;
+    private Double physicalDeterioration;
     private NetworkMaterial networkMaterial;
     private NetworkThermalInsulationMaterial networkThermalInsulationMaterial;
     private RisersMaterial risersMaterial;
